[WOF-US05] code refactor
