+++ src/main/java/cloud/cholewa/App.java
@@ -1,9 +1,18 @@
 package cloud.cholewa;
 
+import cloud.cholewa.exceptions.GameException;
 import cloud.cholewa.game.logic.GameEngine;
+
+import java.util.InputMismatchException;
+
+import static cloud.cholewa.common.ConsoleControl.showConsoleMessage;
 
 public class App {
     public static void main(String[] args) {
-        GameEngine.run();
+        try {
+            GameEngine.run();
+        } catch (GameException | IllegalArgumentException e) {
+            showConsoleMessage(e.getMessage());
+        }
     }
 }
